requirements.txt
fastapi
uvicorn
asterisk-ami
requests
python-dotenv
elevenlabs
pydub
pygame
pyaudio
wave
